Merge20220202 1. Correct all SAL warnings. 2. Correct typo in name callback in DMF_NotifyUserWithRequest. 3.  Correct error DMF_String. Validation of buffer length for two Methods was incorrect.

+++ Dmf/Modules.Library/Dmf_HashTable.c
@@ -154,7 +154,7 @@ Return Value:
 
 --*/
 {
-    return (DATA_ENTRY*)((UCHAR*)ModuleContext->DataTable + ModuleContext->DataEntrySize * EntryIndex);
+    return (DATA_ENTRY*)((UCHAR*)ModuleContext->DataTable + (size_t)ModuleContext->DataEntrySize * (size_t)EntryIndex);
 }
 
 static
@@ -400,7 +400,7 @@ Return Value:
                   sizeToAllocate,
                   INVALID_INDEX);
 
-    sizeToAllocate = moduleContext->DataTableSize * moduleContext->DataEntrySize;
+    sizeToAllocate = (size_t)moduleContext->DataTableSize * (size_t)moduleContext->DataEntrySize;
     DmfAssert(sizeToAllocate != 0);
 
     WDF_OBJECT_ATTRIBUTES_INIT(&objectAttributes);
